mrc-load-model-utils(extract-downgrade-dtype-move,mechanical_provable): move maybe_downgrade_dtype_for_legacy_gpu to load_model_utils module (cut+paste)
